# backend/openapi/v1.yaml
openapi: 3.0.3
info:
  title: DataAfrik API
  version: 1.0.0
  description: >
    API-first contract for DataAfrik domain services. This v1 spec is a migration
    baseline and is expanded incrementally while legacy endpoints remain active.
servers:
  - url: /api/v1
...
paths:
...
  /comments:
    post:
      tags: [comments]
      summary: Create comment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Comment created
  /reports/request:
    post:
      tags: [reports]
      summary: Submit custom report request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Request accepted
...
  /subscriptions/pdf-checkout:
    post:
      tags: [subscriptions]
      summary: Start PDF checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Checkout initialized
...
  /subscriptions/newsletter:
    post:
      tags: [subscriptions]
      summary: Subscribe email to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Subscription result
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...